Aggregate parser files
